Refactored code to make classes shorter
