+++ Phase0/src/worksheet_maker/EquationsToPDImages.java
@@ -1,8 +1,11 @@
 package worksheet_maker;
 
+import org.apache.pdfbox.pdmodel.PDDocument;
+import org.apache.pdfbox.pdmodel.graphics.image.PDImageXObject;
 import org.scilab.forge.jlatexmath.TeXFormula;
 
 import java.awt.image.BufferedImage;
+import java.io.IOException;
 import java.util.Map;
 
 /**
@@ -12,7 +15,7 @@ import java.util.Map;
  * @version 1.0
  * @since 2021-11-5.
  */
-public class EquationsToResizedImages {
+public class EquationsToPDImages {
     private final EquationStringToLatex equationStringToLatex = new EquationStringToLatex();
     private final LatexToImage latexToImage = new LatexToImage();
 
@@ -25,13 +28,13 @@ public class EquationsToResizedImages {
      * @return A list of resized image lists. The first array represents the images for the questions, the second array
      * represents the images for the answers.
      */
-    public BufferedImage[][] createResizedImages(Map<String, Object> formatEquationDetails, String[][] equations) {
-        BufferedImage[][] qAndAImages = new BufferedImage[2][equations.length];
+    public PDImageXObject[][] createResizedImages(Map<String, Object> formatEquationDetails, String[][] equations, PDDocument[] worksheetPDFs) throws IOException {
+        PDImageXObject[][] qAndAImages = new PDImageXObject[2][equations.length];
         for (int i = 0; i < equations.length; i++) {
             for (int ans = 0; ans < 2; ans++) { //If ans = 1, add it to the answer. Otherwise, add it to question.
                 TeXFormula questionFormula = equationStringToLatex.convertEquationStringToLatex(equations[i],
                         (String) formatEquationDetails.get("equationFormat"), ans == 1);
-                qAndAImages[ans][i] = latexToImage.convertLatexToImage((new TeXFormula((i+1) + ")\\;\\;\\;")).add(questionFormula));
+                qAndAImages[ans][i] = latexToImage.convertLatexToImage((new TeXFormula((i+1) + ")\\;\\;\\;")).add(questionFormula), worksheetPDFs[ans]);
             }
         }
         return qAndAImages;
